[main] Seeding Service

+++ WebSales/Models/Seller.cs
@@ -10,7 +10,7 @@ namespace WebSales.Models
         public string Name { get; set; }
         public string Email { get; set; }
         public DateTime BirthDate { get; set; }
-        public int BaseSalary{ get; set; }
+        public double BaseSalary{ get; set; }
         public Department Department { get; set; }
         public ICollection<SalesRecord> Sales { get; set; } = new List<SalesRecord>();
 
@@ -19,7 +19,7 @@ namespace WebSales.Models
 
         }
 
-        public Seller(int id, string name, string email, DateTime birthDate, int baseSalary, Department department)
+        public Seller(int id, string name, string email, DateTime birthDate, double baseSalary, Department department)
         {
             Id = id;
             Name = name;
